fixup! Initial draft of the guts of the adaptive padding machine.
I had to implement real machine creation events and generalize the machine
handling for a few reasons:
  1. Without machine creation events, nothing made sense. Without this,
     we'd just be hard coding machine creation in the middle of relay.c,
     which was wrong and could not be merged.
  2. Torrc serialization wouldn't be possible without this code
  3. It helped find bugs. For example, I didn't realize we needed an ack on shutdown
     to properly recognize padding packets as valid until I did this.
  4. I also realized that we needed to specify which machine index to use,
     or machine interaction event ordering would be non-deterministic.

Now all of the padding creation is as it should be, and supports different
types of padding for different types of circuits.

Plus, all of this took only two days, not 3 weeks. Sorry, not sorry.