fixup! Initial draft of the guts of the adaptive padding machine.
Implement length-based termination and iat dist sampling as an alternative to
histograms.

We may have lost the WTF-PAD histograms and they might not apply anyway. This
way we can at least specify APE or the distributions from WTF-PAD's
simulators.

We only provide logistic and log-logistic because they are inveritble in
closed form. lognormal is not.