Add events for negotiation, and dummy machine to gather vanilla traces.
This adds four new events, two on the client and two on the relay.

Client side:
  - circpad_negotiate_padding
  - circpad_handle_padding_negotiated

Relay side:
  - circpad_handle_padding_negotiate
  - circpad_padding_negotiated

The simulator's log extraction python script should *always* strip these out.

For trace gathering, the dummy machine will always be negotiated as soon as
the second hop is built, and should last until the circuit is closed.

The ordering for this, then is:

Client:
  1. circpad_negotiate_padding
  2. circpad_cell_event_nonpadding_sent
  3. circpad_cell_event_nonpadding_received
  4. circpad_handle_padding_negotiated

Relay:
  1. circpad_handle_padding_negotiate
  2. circpad_padding_negotiated
  3. circpad_cell_event_nonpadding_sent

The log extraction script should have an option to also remove all of these
events from both relay and client, so that gathered vanilla traces are not
polluted by the extra negotiate cell events.

Note that the relay side is missing the first circpad_cell_event_nonpadding_received.

It won't be present on there, *unless* there is already a machine opened on
the circuit... Which can happen in some cases..  bleh.