Sim: Specify circid for relay + client.
Made client circid a mandatory arg for custom traces to allow experiments that
involve circuit multiplexing (this way circids are preserved and traces can be
split+merged).

Maybe we want to put this circid in the python generated trace format instead
to avoid confusion/trace mixup. But this was a simpler change.